feat: import .csv file and tests

+++ entrega1/src/main/java/com/tingeso/entrega1/entities/Cuota.java
@@ -16,7 +16,7 @@ public class Cuota {
     private Long idCuota;
     private Integer numeroCuota;
     private Integer monto;
-    private Integer descuento;
+    private float descuento;
     private int mesesAtraso;
     private float interes;
     private Boolean pagado;
